views_v2 and serializers_v2
Refactored awards package to include views_v2 and serializers_v2. Combined award_type and recipient py files into award_spending.
