+++ usaspending_api/awards/serializers.py
@@ -10,14 +10,6 @@ from usaspending_api.references.serializers import AgencySerializer, LegalEntity
 from usaspending_api.common.helpers import fy
 
 
-class AwardTypeAwardSpendingSerializer(serializers.Serializer):
-
-    award_type = serializers.CharField()
-    toptier_agency = serializers.CharField()
-    subtier_agency = serializers.CharField()
-    obligated_amount = serializers.DecimalField(None, 2)
-
-
 class FinancialAccountsByAwardsSerializer(LimitableSerializer):
 
     class Meta:
@@ -64,19 +56,6 @@ class FinancialAccountsByAwardsSerializer(LimitableSerializer):
                 "kwargs": {"read_only": True}
             },
         }
-
-
-class RecipientSeriallizer(serializers.Serializer):
-    recipient_id = serializers.IntegerField()
-    recipient_name = serializers.CharField()
-
-
-class RecipientAwardSpendingSerializer(serializers.Serializer):
-
-    recipient = RecipientSeriallizer(source='*')
-    toptier_agency = serializers.CharField()
-    subtier_agency = serializers.CharField()
-    obligated_amount = serializers.DecimalField(None, 2)
 
 
 class SubawardSerializer(LimitableSerializer):
