+++ usaspending_api/awards/urls_award_spending.py
@@ -1,11 +1,10 @@
 from django.conf.urls import url
 
-from usaspending_api.awards import award_type as type_views
-from usaspending_api.awards import recipient as recipient_views
+from usaspending_api.awards.views_v2 import award_spending as views
 
 # map reqest types to viewset method; replace this with a router
-award_type = type_views.AwardTypeAwardSpendingViewSet.as_view({'get': 'list'})
-recipient = recipient_views.RecipientAwardSpendingViewSet.as_view({'get': 'list'})
+award_type = views.AwardTypeAwardSpendingViewSet.as_view({'get': 'list'})
+recipient = views.RecipientAwardSpendingViewSet.as_view({'get': 'list'})
 
 urlpatterns = [
     url(r'^award_type/', award_type, name='award-type-award-spending'),
